fix: 소켓 연결 시점 변경
- 기존 구현 사항으로 닉네임을 입력 받아 해당 닉네임을 소켓의 connect 이벤트에서 서버에 전달할 예정이었습니다. 하지만 server.js에서 초기화한 소켓 인스턴스가 원하는 시점에 초기화되지 않고 이미 초기화된 상태로 컴포넌트가 렌더링 되어 connect 이벤트가 발생하지 않았습니다.
- 이에 socket의 autoConnect를 false로 설정한 뒤 닉네임 입력 후 submit 이벤트가 발생할 때 소켓을 연결할 수 있도록 설정했습니다.
