--- src/components/Socket.jsx
@@ -1,40 +1,0 @@
-import React from "react";
-import { useEffect, useState } from "react";
-import { socket } from "./../utils/server";
-
-export default function Socket() {
-  const [isConnected, setIsConnected] = useState(false);
-  const [transport, setTransport] = useState("N/A");
-
-  useEffect(() => {
-    const onConnect = () => {
-      setIsConnected(true);
-      setTransport(socket.io.engine.transport.name);
-
-      socket.io.engine.on("upgrade", (transport) => {
-        setTransport(transport.name);
-      });
-    };
-
-    const onDisconnect = () => {
-      setIsConnected(false);
-      setTransport("N/A");
-    };
-
-    socket.on("connect", onConnect);
-    socket.on("disconnect", onDisconnect);
-
-    return () => {
-      socket.off("connect", onConnect);
-      socket.off("disconnect", onDisconnect);
-    };
-  }, []);
-  return (
-    <section>
-      <article>
-        <p>Status: {isConnected ? "connected" : "disconnected"}</p>
-        <p>Transport: {transport}</p>
-      </article>
-    </section>
-  );
-}
