+++ src/utils/server.js
@@ -2,4 +2,6 @@ import { io } from "socket.io-client";
 
 const URL = "http://localhost:4000";
 
-export const socket = io(URL);
+export const socket = io(URL, {
+  autoConnect: false,
+});
